Move tests to AlterTablePartitionParserSuite
